+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version 0.18.{1,2} (2019-12-09)
 - bugfix: cost analyze: use resample instead of manually constructing the binned dataframe with `date_range`
   - found this issue when testing `isitfit cost analyze --ndays=15`
 - bugfix: cost analyze: when no ec2/redshift data found, just raise an exception in the binning aggregator and abort early
+- bugfix: cost optimize: skip the step of binning for redshift cost optimize in the "common" pipeline factory
 
 
 Version 0.18.0 (2019-12-09)