+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Version 0.18.{1,2} (2019-12-09)
 - bugfix: cost analyze: when no ec2/redshift data found, just raise an exception in the binning aggregator and abort early
 - bugfix: cost optimize: skip the step of binning for redshift cost optimize in the "common" pipeline factory
 - enh: cost analyze: pretty-print dollars/percentages in binned report
+- feat: cost: increase `config.retries.max_attempts` from 4 to 10 (knowing that botocore has exponential backoff) to get around rate limiting
 
 
 Version 0.18.0 (2019-12-09)