+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version latest (0.18.0rc?, 2019-12-05?)
 - ...
 
 
-Version 0.18.{1,2} (2019-12-09)
+Version 0.18.{1,2,3} (2019-12-09)
 
 - bugfix: cost analyze: use resample instead of manually constructing the binned dataframe with `date_range`
   - found this issue when testing `isitfit cost analyze --ndays=15`